restructure folders based on react-startkit standards
