--- src/app/theme/_config.scss
@@ -1,4 +1,0 @@
-@import "~react-toolbox/lib/colors";
-
-$color-primary: $palette-blue-500;
-$color-primary-dark: $palette-blue-700;
