+++ src/components/AppBar/AppBar.jsx
@@ -1,11 +1,11 @@
 import React, { PropTypes } from 'react';
 import { AppBar } from 'react-toolbox/lib/app_bar';
-import Logo from './Logo.jsx';
-import theme from './PurpleAppBar.scss';
+import Logo from './../Logo/Logo.jsx';
+import theme from './AppBar.scss';
 
 const PurpleAppBar = ({ children, ...other }) => (
   <AppBar {...other} theme={theme}>
-    <Logo /> App Example
+    <Logo />Splitspense
     {children}
   </AppBar>
 );
