+++ src/components/Button/SuccessButton.scss
