+++ src/theme/_config.scss
@@ -1,0 +1,4 @@
+@import "~react-toolbox/lib/colors";
+
+$color-primary: $color-black;
+$color-primary-dark: $palette-green-900;
