+++ src/www/index.html
@@ -2,9 +2,9 @@
 <html lang="en">
   <head>
     <meta charset="utf-8">
-    <title>React Toolbox Example</title>
-    <meta name="description" content="A set of complementary tools to ReactJS.">
-    <meta name="author" content="React Toolbox team">
+    <title>Splitspense</title>
+    <meta name="description" content="Splitting bills among friends made easy">
+    <meta name="author" content="Leanne Zhang">
     <meta name="viewport" content="initial-scale=1.0,user-scalable=no,maximum-scale=1,width=device-width">
     <meta name="viewport" content="initial-scale=1.0,user-scalable=no,maximum-scale=1" media="(device-height: 568px)">
     <meta name="apple-mobile-web-app-title" content="Material Console">
